ContextMenu
Create base class for menu entries
Move menu entries realizations into new files
Move files in "context_menu" folder
